requirements.txt
google_api_python_client==2.97.0
google_auth_oauthlib==1.0.0
openai==0.27.6
protobuf==3.20.0
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.26.0
tweepy==4.14.0
schedule
wheel
mysql-connector-python==9.0.0
stripe==11.4.1